requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py311haa95532_0
filelock=3.10.7=pypi_0
imageio=2.27.0=pypi_0
jinja2=3.1.2=pypi_0
lazy-loader=0.2=pypi_0
libffi=3.4.2=hd77b12b_6
markupsafe=2.1.2=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.0=pypi_0
numpy=1.24.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=pypi_0
pillow=9.4.0=pypi_0
pip=23.0.1=py311haa95532_0
python=3.11.2=h966fe2a_0
pywavelets=1.4.1=pypi_0
rawpy=0.18.0=pypi_0
scikit-image=0.20.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=65.6.3=py311haa95532_0
sqlite=3.41.1=h2bbff1b_0
sympy=1.11.1=pypi_0
tifffile=2023.3.21=pypi_0
tk=8.6.12=h2bbff1b_0
torch=2.0.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2022g=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
wincertstore=0.2=py311haa95532_0
xz=5.2.10=h8cc25b3_1
zlib=1.2.13=h8cc25b3_0
